First working version
This version is now working and has the correct spelling of the header
needed to detect Tactile.  It will serve whatever the default theme is
set to for desktop browsers coming in to the WordPress site URL, and
for browsers arriving through the Netbiscuits front end URL it will
serve a theme called twentyten for devices that can support Tactile
markup and a theme called christmas-is-near for devices that can only
support BiscuitML markup.  Obviously these themes need to be installed
and should also output the correct markup.  This plugin does not come
with themes you will need to make your own.